value_list: Migrate functions related to `value_list_t` out of `common`.
