+++ src/utils/value_list/value_list.h
@@ -29,6 +29,9 @@
 #ifndef UTILS_VALUE_LIST_H
 #define UTILS_VALUE_LIST_H 1
 
+#include "daemon/data_set.h"
+#include "daemon/plugin.h"
+
 struct value_list_s {
   value_t *values;
   size_t values_len;
@@ -96,5 +99,33 @@ int plugin_dispatch_values(value_list_t const *vl);
 __attribute__((sentinel)) int plugin_dispatch_multivalue(value_list_t const *vl,
                                                          bool store_percentage,
                                                          int store_type, ...);
+
+int parse_identifier(char *str, char **ret_host, char **ret_plugin,
+                     char **ret_type, char **ret_data_source,
+                     char *default_host);
+
+/* parse_identifier_vl parses an identifier in the form
+ * "host/plugin[-inst]/type[-inst]/data_source" and stores the fields in a
+ * value_list_t. If vl->host is not empty, then it is used as the default value
+ * if a host name is omitted, i.e. the "plugin/type" format is accepted. If
+ * ret_data_source is not NULL, a four-part identifier is accepted and a
+ * pointer to the data source name is (optionally) stored and needs to be freed
+ * by the caller. If the provided format does not fit the provided arguments,
+ * e.g. a two-part format but no default host provided, or a four-part format
+ * but no ret_data_source pointer, then EINVAL is returned.
+ */
+int parse_identifier_vl(const char *str, value_list_t *vl,
+                        char **ret_data_source);
+
+/* parse_legacy_identifier parses a legacy identifier in the form
+ * "host/plugin/type" and converts it to a metric_t. */
+metric_t *parse_legacy_identifier(char const *s);
+
+/* plugin_value_list_to_metric_family converts a value in a value_list_t to a
+ * metric_family_t. In case of error, errno is set and NULL is returned. The
+ * returned pointer must be freed using metric_family_free(). */
+metric_family_t *plugin_value_list_to_metric_family(value_list_t const *vl,
+                                                    data_set_t const *ds,
+                                                    size_t index);
 
 #endif
